+++ src/game/tnt/TNTEntity.java
@@ -163,7 +163,7 @@ public class TNTEntity {
         indices.add(0); indices.add(1+indicesCount); indices.add(2+indicesCount); indices.add(0); indices.add(2+indicesCount); indices.add(3+indicesCount);
         indicesCount += 4;
 
-        float[] textureFront = getFrontTexturePoints(6,(byte) 0);
+        float[] textureFront = getFrontTexturePoints((byte)6,(byte) 0);
         //front
         textureCoord.add(textureFront[1]);textureCoord.add(textureFront[2]);
         textureCoord.add(textureFront[0]);textureCoord.add(textureFront[2]);
@@ -185,7 +185,7 @@ public class TNTEntity {
         indices.add(indicesCount); indices.add(1+indicesCount); indices.add(2+indicesCount); indices.add(indicesCount); indices.add(2+indicesCount); indices.add(3+indicesCount);
         indicesCount += 4;
 
-        float[] textureBack = getBackTexturePoints(6,(byte) 0);
+        float[] textureBack = getBackTexturePoints((byte) 6,(byte) 0);
         //back
         textureCoord.add(textureBack[1]);textureCoord.add(textureBack[2]);
         textureCoord.add(textureBack[0]);textureCoord.add(textureBack[2]);
@@ -206,7 +206,7 @@ public class TNTEntity {
         indices.add(indicesCount); indices.add(1+indicesCount); indices.add(2+indicesCount); indices.add(indicesCount); indices.add(2+indicesCount); indices.add(3+indicesCount);
         indicesCount += 4;
 
-        float[] textureRight = getRightTexturePoints(6,(byte) 0);
+        float[] textureRight = getRightTexturePoints((byte) 6,(byte) 0);
         //right
         textureCoord.add(textureRight[1]);textureCoord.add(textureRight[2]);
         textureCoord.add(textureRight[0]);textureCoord.add(textureRight[2]);
@@ -227,7 +227,7 @@ public class TNTEntity {
         indices.add(indicesCount); indices.add(1+indicesCount); indices.add(2+indicesCount); indices.add(indicesCount); indices.add(2+indicesCount); indices.add(3+indicesCount);
         indicesCount += 4;
 
-        float[] textureLeft = getLeftTexturePoints(6,(byte) 0);
+        float[] textureLeft = getLeftTexturePoints((byte) 6,(byte) 0);
         //left
         textureCoord.add(textureLeft[1]);textureCoord.add(textureLeft[2]);
         textureCoord.add(textureLeft[0]);textureCoord.add(textureLeft[2]);
@@ -248,7 +248,7 @@ public class TNTEntity {
         indices.add(indicesCount); indices.add(1+indicesCount); indices.add(2+indicesCount); indices.add(indicesCount); indices.add(2+indicesCount); indices.add(3+indicesCount);
         indicesCount += 4;
 
-        float[] textureTop = getTopTexturePoints(6);
+        float[] textureTop = getTopTexturePoints((byte) 6);
         //top
         textureCoord.add(textureTop[1]);textureCoord.add(textureTop[2]);
         textureCoord.add(textureTop[0]);textureCoord.add(textureTop[2]);
@@ -269,7 +269,7 @@ public class TNTEntity {
         //bottom
         indices.add(indicesCount); indices.add(1+indicesCount); indices.add(2+indicesCount); indices.add(indicesCount); indices.add(2+indicesCount); indices.add(3+indicesCount);
 
-        float[] textureBottom = getBottomTexturePoints(6);
+        float[] textureBottom = getBottomTexturePoints((byte) 6);
         //bottom
         textureCoord.add(textureBottom[1]);textureCoord.add(textureBottom[2]);
         textureCoord.add(textureBottom[0]);textureCoord.add(textureBottom[2]);
